:recycle: 포트폴리오 -> about
