+++ src/apps/about/components/about-card.tsx
@@ -67,9 +67,9 @@ const AboutCard = () => {
       </div>
 
       <div className="flex flex-col justify-center flex-1">
-        <div className="w-[80%] mx-auto mt-[-24px] flex items-center justify-center font-bold tracking-tight gap-8pxr">
-          <p className="text-violet text-40pxr tablet:text-[4vw]">front-end</p>
-          <p className="text-primary text-48pxr tablet:text-[6vw]">web developer</p>
+        <div className="flex items-center justify-center w-4/5 mx-auto font-extrabold tracking-tight text-transparent -mt-24pxr gap-8pxr from-violet via-pink-600 to-primary bg-gradient-to-r bg-clip-text">
+          <p className="text-40pxr tablet:text-[4vw]">Front-end</p>
+          <p className="text-48pxr tablet:text-[6vw]">web Developer</p>
         </div>
       </div>
